refactor: extract component to separate file
